requirements.txt
Flask>=2.2
flask-cors>=4.0

aiosmtpd>=1.4
aiosmtplib>=3.0
aiohttp>=3.9

python-dotenv>=1.0

PyMySQL==1.1.0
DBUtils>=3.0

chromadb>=0.4.22
sentence-transformers>=2.6
pandas>=2.0

beautifulsoup4>=4.12
cryptography>=42.0

gunicorn>=21.2
gevent>=24.2

celery>=5.3
redis>=5.0

pycryptodome==3.19.0